indexed mesh: Fix setting attribute offsets
Offset of vertex attribute pointer function works when there is a buffer
attached to GL_ARRAY_BUFFER. This was not the case and is not corrected.

Create info had redundant information:
Vertex size and vertex count can be gotten from attributes vector and
the total size of the vertex buffer which is more important to know.
Index size and count, similarly can be gotten from the total index buffer
size as the only sizes of indices we support as of yet are 16 bit.